+++ aiohttp/client.py
@@ -66,12 +66,9 @@ from .client_reqrep import ClientRequest as ClientRequest
 from .client_reqrep import ClientResponse as ClientResponse
 from .client_reqrep import Fingerprint as Fingerprint
 from .client_reqrep import RequestInfo as RequestInfo
-from .client_utils import (
-    DEFAULT_TIMEOUT,
-    DEFAULT_WS_CLIENT_TIMEOUT,
-    ClientTimeout,
-)
+from .client_ws import DEFAULT_WS_CLIENT_TIMEOUT
 from .client_ws import ClientWebSocketResponse as ClientWebSocketResponse
+from .client_ws import ClientWSTimeout
 from .connector import BaseConnector as BaseConnector
 from .connector import NamedPipeConnector as NamedPipeConnector
 from .connector import TCPConnector as TCPConnector
@@ -136,11 +133,35 @@ __all__ = (
     'ClientTimeout',
     'request')
 
-
 try:
     from ssl import SSLContext
 except ImportError:  # pragma: no cover
     SSLContext = object  # type: ignore
+
+
+@attr.s(frozen=True, slots=True)
+class ClientTimeout:
+    total = attr.ib(type=Optional[float], default=None)
+    connect = attr.ib(type=Optional[float], default=None)
+    sock_read = attr.ib(type=Optional[float], default=None)
+    sock_connect = attr.ib(type=Optional[float], default=None)
+
+    # pool_queue_timeout = attr.ib(type=float, default=None)
+    # dns_resolution_timeout = attr.ib(type=float, default=None)
+    # socket_connect_timeout = attr.ib(type=float, default=None)
+    # connection_acquiring_timeout = attr.ib(type=float, default=None)
+    # new_connection_timeout = attr.ib(type=float, default=None)
+    # http_header_timeout = attr.ib(type=float, default=None)
+    # response_body_timeout = attr.ib(type=float, default=None)
+
+    # to create a timeout specific for a single request, either
+    # - create a completely new one to overwrite the default
+    # - or use http://www.attrs.org/en/stable/api.html#attr.evolve
+    # to overwrite the defaults
+
+
+# 5 Minute default read timeout
+DEFAULT_TIMEOUT = ClientTimeout(total=5*60)
 
 
 _RetType = TypeVar('_RetType')
@@ -565,7 +586,7 @@ class ClientSession:
             url: StrOrURL, *,
             method: str=hdrs.METH_GET,
             protocols: Iterable[str]=(),
-            timeout: Union[ClientTimeout, float]=sentinel,
+            timeout: Union[ClientWSTimeout, float]=sentinel,
             receive_timeout: Optional[float]=None,
             autoclose: bool=True,
             autoping: bool=True,
@@ -604,7 +625,7 @@ class ClientSession:
             url: StrOrURL, *,
             method: str=hdrs.METH_GET,
             protocols: Iterable[str]=(),
-            timeout: Union[ClientTimeout, float]=sentinel,
+            timeout: Union[ClientWSTimeout, float]=sentinel,
             receive_timeout: Optional[float]=None,
             autoclose: bool=True,
             autoping: bool=True,
@@ -619,27 +640,25 @@ class ClientSession:
             compress: int=0,
             max_msg_size: int=4*1024*1024
     ) -> ClientWebSocketResponse:
-
-        ws_timeout: ClientTimeout
         if timeout is not sentinel:
-            if isinstance(timeout, ClientTimeout):
+            if isinstance(timeout, ClientWSTimeout):
                 ws_timeout = timeout
             else:
                 warnings.warn("parameter 'timeout' of type 'float' "
                               "is deprecated, please use "
-                              "'timeout=ClientTimeout(sock_close=...)'",
+                              "'timeout=ClientWSTimeout(ws_close=...)'",
                               DeprecationWarning,
                               stacklevel=2)
-                ws_timeout = ClientTimeout(sock_close=timeout)
+                ws_timeout = ClientWSTimeout(ws_close=timeout)
         else:
             ws_timeout = DEFAULT_WS_CLIENT_TIMEOUT
         if receive_timeout is not None:
             warnings.warn("float parameter 'receive_timeout' "
                           "is deprecated, please use parameter "
-                          "'timeout=ClientTimeout(sock_read=...)'",
+                          "'timeout=ClientWSTimeout(ws_receive=...)'",
                           DeprecationWarning,
                           stacklevel=2)
-            ws_timeout = attr.evolve(ws_timeout, sock_read=receive_timeout)
+            ws_timeout = attr.evolve(ws_timeout, ws_receive=receive_timeout)
 
         if headers is None:
             real_headers = CIMultiDict()  # type: CIMultiDict[str]
